+++ templates/part.content.php
@@ -16,68 +16,7 @@
     <h2 id="reportHeader"></h2>
     <h3 id="reportSubHeader" hidden></h3>
     <div id="reportPlaceholder"></div>
-    <div id="reportMenuBar" style="position: relative; height: 24px;">
-        <span id="reportMenuIcon" class="analytics-options icon-more has-tooltip"
-              title="<?php p($l->t('Options')); ?>"></span>
-        <div id="reportMenu" class="popovermenu" style="top: 33px; right: -5px;">
-            <ul id="reportMenuMain">
-                <li id="saveIcon">
-                    <button>
-                        <span class="icon-analytics-save-warning"
-                              title="<?php p($l->t('Report was changed - Press here to save the current state')); ?>"></span>
-                        <span><?php p($l->t('Save view')); ?></span>
-                    </button>
-                </li>
-                <li>
-                    <button id="drilldownIcon">
-                        <span class="icon-analytics-drilldown"></span>
-                        <span><?php p($l->t('Drilldown')); ?></span>
-                    </button>
-                </li>
-                <li>
-                    <button id="chartOptionsIcon">
-                        <span class="icon-analytics-chart-options"></span>
-                        <span><?php p($l->t('Chart options')); ?></span>
-                    </button>
-                </li>
-                <li>
-                    <button id="analysisIcon">
-                        <span class="icon-analytics-forecast"></span>
-                        <span><?php p($l->t('Analysis')); ?></span>
-                    </button>
-                </li>
-                <li>
-                    <button id="downlaodChartIcon">
-                        <span class="icon-download"></span>
-                        <span><?php p($l->t('Download chart')); ?></span>
-                        <a id="downlaodChartLink" href='' download="Chart.png" hidden>-</a>
-                    </button>
-                </li>
-
-            </ul>
-            <ul id="reportMenuAnalysis" style="display: none !important;">
-                <li id="backIcon">
-                    <button>
-                        <span class="icon-view-previous"></span>
-                        <span><?php p($l->t('back')); ?></span>
-                    </button>
-                </li>
-                <li>
-                    <button id="trendIcon">
-                        <span><?php p($l->t('Trend')); ?></span>
-                    </button>
-                </li>
-                <li>
-                    <button id="linearRegressionIcon" class="menuitem" disabled>
-                        <span><?php p($l->t('Linear Regression')); ?></span>
-                    </button>
-                </li>
-            </ul>
-        </div>
-        <div id="addFilterIcon" class="analytics-options icon-analytics-filter-add has-tooltip"
-             title="<?php p($l->t('Filter')); ?>"></div>
-        <div id="filterVisualisation" style="display: inline-block; float: right;"></div>
-    </div>
+    <?php print_unescaped($this->inc('part.menu')); ?>
     <div id="chartContainer">
         <canvas id="myChart"></canvas>
     </div>